requirements.txt
asgiref==3.5.2
attrs==21.4.0
boto3==1.23.6
botocore==1.26.6
Django==4.0.4
django-cors-headers==3.12.0
django-storages==1.12.3
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.12.0
gunicorn==20.1.0
inflection==0.5.1
iniconfig==1.1.1
jmespath==1.0.0
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.1.0
